refactor: add choose-peers folder and create body
