+++ src/routes/sign-up/steps/index.js
@@ -1,4 +1,4 @@
-import { ChoosePeers } from "./choose-peers";
+import { ChoosePeers } from "./choose-peers/choose-peers";
 import { FinishReg } from "./finish-reg";
 import { NotifyPeers } from "./notify-peers";
 import { RepeatSeedPhrase } from "./repeat-seed-phrase/repeat-seed-phrase";
